don't overwrite example.com if it exists
instead we hook into it using the hook_provision_apache_vhost_config()

unfotunately, this will not work right away - the site will need to be
verified.

also, if the subdir site is deleted, the apache config will be
corrupt, we would need to generate (and not verify, because then it's
too late) example.com then too.

finally, that glob() is highly suspicious, and none of this is tested
- i suspect http_subdird_path is not populated here

this is all to fix #2020089